Refactor threejs setup code into separate file
