+++ src/components/CanvasRenderer/CanvasRenderer.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import * as THREE from 'three';
+import setup from '../../3d/setup';
 import './CanvasRenderer.css';
 
 class CanvasRenderer extends React.Component {
@@ -9,25 +9,7 @@ class CanvasRenderer extends React.Component {
   }
 
   componentDidMount() {
-    this.camera = new THREE.PerspectiveCamera(75, window.innerWidth / window.innerHeight, 0.1, 1000);
-    this.camera.position.z = 5;
-    this.scene = new THREE.Scene();
-
-    this.renderer = new THREE.WebGLRenderer({ canvas: this.canvas.current });
-    this.renderer.setSize(window.innerWidth, window.innerHeight);
-
-    const geometry = new THREE.SphereGeometry(1, 1, 1);
-    const material = new THREE.MeshBasicMaterial({ color: 0xffffff });
-    const cube = new THREE.Mesh(geometry, material);
-    this.scene.add(cube);
-
-    const animate = () => {
-      requestAnimationFrame(animate);
-      this.renderer.render(this.scene, this.camera);
-      cube.rotation.x += 0.01;
-      cube.rotation.y += 0.01;
-    };
-    animate();
+    setup(this.canvas.current);
   }
 
   render() {
